refactoring _RelatedAdvanced
