+++ src/components/Related/index.js
@@ -1,4 +1,4 @@
-import React, { PureComponent, useState } from 'react';
+import React, { PureComponent } from 'react';
 import T from 'prop-types';
 import { connect } from 'react-redux';
 import { createSelector } from 'reselect';
@@ -7,17 +7,14 @@ import { omit } from 'lodash-es';
 import Link from 'components/generic/Link';
 import loadData from 'higherOrder/loadData';
 
-import Matches from 'components/Matches';
 import Loading from 'components/SimpleCommonComponents/Loading';
 
 import { toPlural } from 'utils/pages';
 
+import RelatedTable from 'components/Related/RelatedTable';
 import EntriesOnStructure from 'components/Related/DomainEntriesOnStructure';
 import StructureOnProtein from 'components/Related/DomainStructureOnProtein';
-import KeySpeciesTable from 'components/Taxonomy/KeySpeciesTable';
-
-import FiltersPanel from 'components/FiltersPanel';
-import CurationFilter from 'components/Protein/ProteinListFilters/CurationFilter';
+import KeySpeciesArea from 'components/Related/Taxonomy/KeySpecies';
 import { getUrlForMeta, getReversedUrl } from 'higherOrder/loadData/defaults';
 
 import { foundationPartial } from 'styles/foundation';
@@ -112,182 +109,6 @@ const mapStateToPropsSimple = createSelector(
 );
 const RelatedSimple = connect(mapStateToPropsSimple)(_RelatedSimple);
 
-const primariesAndSecondaries = {
-  entry: {
-    protein: {
-      primary: 'protein',
-      secondary: 'entry',
-    },
-    structure: {
-      primary: 'structure',
-      secondary: 'entry',
-    },
-    taxonomy: {
-      primary: 'taxonomy',
-      secondary: 'entry',
-    },
-    proteome: {
-      primary: 'proteome',
-      secondary: 'entry',
-    },
-    set: {
-      primary: 'set',
-      secondary: 'entry',
-    },
-  },
-  protein: {
-    entry: {
-      primary: 'entry',
-      secondary: 'protein',
-    },
-    structure: {
-      primary: 'structure',
-      secondary: 'protein',
-    },
-  },
-  structure: {
-    entry: {
-      primary: 'entry',
-      secondary: 'structure',
-    },
-    protein: {
-      primary: 'protein',
-      secondary: 'structure',
-    },
-  },
-  taxonomy: {
-    entry: {
-      primary: 'entry',
-      secondary: 'taxonomy',
-    },
-    protein: {
-      primary: 'protein',
-      secondary: 'taxonomy',
-    },
-    structure: {
-      primary: 'structure',
-      secondary: 'taxonomy',
-    },
-    proteome: {
-      primary: 'proteome',
-      secondary: 'taxonomy',
-    },
-  },
-  proteome: {
-    entry: {
-      primary: 'entry',
-      secondary: 'proteome',
-    },
-    protein: {
-      primary: 'protein',
-      secondary: 'proteome',
-    },
-    structure: {
-      primary: 'structure',
-      secondary: 'proteome',
-    },
-  },
-  set: {
-    entry: {
-      primary: 'entry',
-      secondary: 'set',
-    },
-    protein: {
-      primary: 'protein',
-      secondary: 'set',
-    },
-    structure: {
-      primary: 'structure',
-      secondary: 'set',
-    },
-    taxonomy: {
-      primary: 'taxonomy',
-      secondary: 'set',
-    },
-    proteome: {
-      primary: 'proteome',
-      secondary: 'set',
-    },
-  },
-};
-const InfoFilters = (
-  {
-    filters,
-    focusType,
-    databases,
-  } /*: {filters: Array<Object>, focusType: string, databases: Object} */,
-) => {
-  if (!filters || filters.length === 0) return null;
-  return (
-    <div className={f('callout', 'info', 'withicon')}>
-      This list shows only:
-      <ul>
-        {filters.map(([endpoint, { db, accession }]) => (
-          <li key={endpoint}>
-            {toPlural(focusType)} associated with{' '}
-            <b>{accession ? endpoint : toPlural(endpoint)}</b>
-            {accession && (
-              <span>
-                {' '}
-                accession <b>{accession}</b>
-              </span>
-            )}
-            {db && (
-              <span>
-                {' '}
-                from the <b>
-                  {(databases[db] && databases[db].name) || db}
-                </b>{' '}
-                database
-              </span>
-            )}
-            .
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
-};
-InfoFilters.propTypes = {
-  databases: T.object.isRequired,
-  focusType: T.string.isRequired,
-  filters: T.array,
-};
-
-const KeySpeciesArea = ({ focusType, showKeySpecies }) => {
-  const [showTaxoInfo, setShowTaxoInfo] = useState(true);
-  if (focusType !== 'taxonomy') return null;
-  return (
-    <>
-      {showTaxoInfo && (
-        <div className={f('callout', 'info', 'withicon')} data-closable>
-          <button
-            className={f('close-button')}
-            aria-label="Close alert"
-            type="button"
-            data-close
-            onClick={() => setShowTaxoInfo(false)}
-          >
-            <span aria-hidden="true">&times;</span>
-          </button>
-          <h5>
-            The taxonomy information is available for both Key species and all
-            organisms. The below tables are shown based on the preference in
-            InterPro settings. If you wish to change it, please do in the{' '}
-            <Link to={{ description: { other: ['settings'] } }}>Settings</Link>{' '}
-            page
-          </h5>
-        </div>
-      )}
-      {showKeySpecies && <KeySpeciesTable />}
-    </>
-  );
-};
-
-KeySpeciesArea.propTypes = {
-  focusType: T.string,
-  showKeySpecies: T.bool,
-};
 /*:: type relatedAdvancedProps = {
   mainData: Object,
   secondaryData: Array<Object>,
@@ -325,7 +146,6 @@ export class _RelatedAdvanced extends PureComponent /*:: <relatedAdvancedProps> 
     showAllSpecies: T.bool.isRequired,
   };
 
-  // eslint-disable-next-line complexity
   render() {
     const {
       mainData,
@@ -340,13 +160,6 @@ export class _RelatedAdvanced extends PureComponent /*:: <relatedAdvancedProps> 
       showKeySpecies,
       showAllSpecies,
     } = this.props;
-    const databases =
-      (dataBase &&
-        !dataBase.loading &&
-        dataBase.payload &&
-        dataBase.payload.databases) ||
-      {};
-    const hasFilters = focusType === 'protein';
     return (
       <div className={f('row', 'column')}>
         <KeySpeciesArea focusType={focusType} showKeySpecies={showKeySpecies} />
@@ -366,65 +179,17 @@ export class _RelatedAdvanced extends PureComponent /*:: <relatedAdvancedProps> 
             ) : null}
             {(focusType === 'taxonomy' && showAllSpecies) ||
             focusType !== 'taxonomy' ? (
-              <>
-                <p>
-                  This {mainType} matches
-                  {secondaryData.length > 1
-                    ? ` these ${toPlural(focusType)}:`
-                    : ` this ${focusType}:`}
-                </p>
-                <InfoFilters
-                  filters={otherFilters}
-                  focusType={focusType}
-                  databases={databases}
-                />
-                <div className={f('row')}>
-                  {hasFilters && (
-                    <div
-                      className={f(
-                        'columns',
-                        'small-12',
-                        'medium-3',
-                        'large-2',
-                        'no-padding',
-                      )}
-                    >
-                      <div className={f('browse-side-panel')}>
-                        <FiltersPanel>
-                          <CurationFilter label="UniProt Curation" />
-                        </FiltersPanel>
-                      </div>
-                    </div>
-                  )}
-                  <div
-                    className={f(
-                      'columns',
-                      'small-12',
-                      hasFilters ? 'medium-9' : 'medium-12',
-                      hasFilters ? 'large-10' : 'large-12',
-                    )}
-                  >
-                    <Matches
-                      {...this.props}
-                      actualSize={actualSize}
-                      matches={secondaryData.reduce(
-                        (prev, { coordinates, ...secondaryData }) => [
-                          ...prev,
-                          {
-                            [mainType]: mainData,
-                            [focusType]: secondaryData,
-                            coordinates,
-                          },
-                        ],
-                        [],
-                      )}
-                      isStale={isStale}
-                      databases={databases}
-                      {...primariesAndSecondaries[mainType][focusType]}
-                    />
-                  </div>
-                </div>
-              </>
+              <RelatedTable
+                mainType={mainType}
+                mainData={mainData}
+                secondaryData={secondaryData}
+                focusType={focusType}
+                otherFilters={otherFilters}
+                dataBase={dataBase}
+                isStale={isStale}
+                actualSize={actualSize}
+                otherProps={{ ...this.props }}
+              />
             ) : null}
           </div>
         )}
@@ -539,13 +304,8 @@ class Related extends PureComponent /*:: <RelatedProps> */ {
   };
 
   render() {
-    const {
-      data,
-      focusType,
-      hasSecondary,
-      showAllSpecies,
-      ...props
-    } = this.props;
+    const { data, focusType, hasSecondary, showAllSpecies, ...props } =
+      this.props;
     if (data.loading) return <Loading />;
     let RelatedComponent = RelatedSimple;
     if (hasSecondary) {
